​​refactor: 도메인 관련 api 타입 위치 이동
