+++ src/apis/cartItems/addCartItems.ts
@@ -1,4 +1,4 @@
-import { AddCartItems } from "@/types/cartItem";
+import { AddCartItems } from "@/apis/cartItems/cartItem.type";
 import { httpClient } from "../httpClient";
 
 const ERROR_MESSAGE = "장바구니에 상품을 추가하던 중 에러가 발생했습니다.";
