+++ src/apis/cartItems/getCartItems.ts
@@ -1,4 +1,4 @@
-import { CartItemType } from "@/types/cartItem";
+import { CartItemType } from "@/apis/cartItems/cartItem.type";
 import { httpClient } from "../httpClient";
 
 const ERROR_MESSAGE = "징바구니를 가져오는 데 실패했습니다.";
